% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{add_previous_item}
\alias{add_previous_item}
\title{Add id of previous item}
\usage{
add_previous_item(
  data,
  col_item = item,
  col_parent = item_parent,
  col_order = item
)
}
\description{
Add id of previous item
}
