% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue.R
\name{get_glue}
\alias{get_glue}
\title{Extract the items that glue together the features. The result will contain an additional
column with the item in which always two features occur.}
\usage{
get_glue(pairs, items)
}
\arguments{
\item{pairs}{A data frame with cooccurrences or sequences, e.g. the result of
resample_cooccurrence, with at least the following columns:
- source The source feature
- target The target feature}

\item{data}{A data frame containing the items with their features, e.g. the input tu resample_cooccurrence,
with at least the following columns:
- item contains the ID of a document (e.g. a comment)
- feature contains a category  (e.g. the topic of the comment)}
}
\value{
A tibble
}
\description{
Extract the items that glue together the features. The result will contain an additional
column with the item in which always two features occur.
}
