% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{get_sequences}
\alias{get_sequences}
\title{Count sequences and calculate probabilities and npmi}
\usage{
get_sequences(data, .progress = NULL)
}
\arguments{
\item{data}{A data frame containing the columns item, feature and optionally weight.
- item contains the ID of a document (e.g. a comment)
_ item_prev contains the ID of a preceding document (e.g. a post)
- feature contains a category  (e.g. the topic of the comment)
- weight  The weight of the feature ranges from 0 to 1 (e.g. how prevalent is the topic in the comment?). Will be initialized with 1 if missing.
Items with multiple features should occur multiple times in the data table.
Make sure not to include duplicates (that's where the weight kicks in, use it!).
The feature and item columns can contain numeric IDs, character strings or factors.
Provide character strings or factors to complete combinations not found in the data.}
}
\value{
A tibble
}
\description{
Count sequences and calculate probabilities and npmi
}
