% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixmap.R
\name{matrixmap}
\alias{matrixmap}
\title{Creates a heatmap}
\usage{
matrixmap(
  data,
  caption = "",
  value.min = NULL,
  value.max = NULL,
  value.mid = NULL
)
}
\arguments{
\item{data}{A tibble with columns source, target, target_label,
source_group, target_group, value}

\item{caption}{The legend caption}
}
\description{
Creates a heatmap
}
