% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/network.R
\name{network}
\alias{network}
\title{Plots a network}
\usage{
network(data)
}
\arguments{
\item{data}{A tibble with columns
source, target,
source_label, target_label,
source_group, target_group,
value, sig}
}
\description{
Plots a network
}
