% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npmi.R
\name{npmi}
\alias{npmi}
\title{Calculate ratio of probabilities and npmi}
\usage{
npmi(data, p_actual, p_expected, smoothing = 0)
}
\arguments{
\item{data}{Data}

\item{p_actual}{Name of column ame containing observed p}

\item{p_expected}{Name of column containing expected p}

\item{smoothing}{Add pseudocount. Calculate the pseudocount based on the number of trials
to apply Laplace's rule of succession.}
}
\description{
Calculate ratio of probabilities and npmi
}
