% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequences.R
\name{resample_sequences}
\alias{resample_sequences}
\title{Calculate npmi and probabilities on the basis of resampling}
\usage{
resample_sequences(data, trials = 10000, smoothing = 0, sig.level = 0.05)
}
\arguments{
\item{data}{A data frame containing the columns item, feature and optionally weight.
- item contains the ID of a document (e.g. a comment)
- feature contains a category  (e.g. the topic of the comment)
- weight  The weight of the feature ranges from 0 to 1 (e.g. how prevalent is the topic in the comment?). Will be initialized with 1 if missing.
Items with multiple features should occur multiple times in the data table.
Make sure not to include duplicates (that's where the weight kicks in, use it!).
The feature and item columns can contain numeric IDs, character strings or factors.
Provide character strings or factors to complete combinations not found in the data.}

\item{trials}{Number of resamples}

\item{smoothing}{Set smoothing to 1 to apply Laplace's rule of succession}
}
\value{
A tibble containing resampleped (random permutation test) cooccurrence metrics.
        Resampled metrics are to be found in columns with the boot prefix.
        Use boot.npmi for the p_cond_source metric to find extraordinary cooccurrence
}
\description{
Calculate npmi and probabilities on the basis of resampling
}
