% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_pairs}
\alias{count_pairs}
\title{Count cooccurrence of features in items}
\usage{
count_pairs(data)
}
\arguments{
\item{data}{A data frame containing the columns item, feature and optionally weight.
- item contains the ID of a document (e.g. a comment)
- feature contains a category  (e.g. the topic of the comment)
- weight  The weight of the feature ranges from 0 to 1 (e.g. how prevalent is the topic in the comment?).
- weight will be initialized with 1 if missing.
Items with multiple features should occur multiple times in the data table.
Make sure not to include duplicates (that's where the weight kicks in, use it!).
The feature and item columns can contain numeric IDs, character strings or factors.
Provide character strings or factors to complete combinations not found in the data.}
}
\value{
A tibble
}
\description{
Count cooccurrence of features in items
}
