% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_pairs}
\alias{count_pairs}
\title{Count cooccurrence of features in items}
\usage{
count_pairs(data, upper = F, diag = F)
}
\arguments{
\item{data}{Data table containin the columns item, feature and optionally weight.
The weight column will be initialized with 1 if missing.}
}
\value{
A data table
}
\description{
Count cooccurrence of features in items
}
