% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count.R
\name{count_sequences}
\alias{count_sequences}
\title{Count sequences of features from sequences of items}
\usage{
count_sequences(data)
}
\arguments{
\item{data}{A data.frame containing the columns item, item_prev, feature.
- item contain the ID of a document (e.g. a comment)
- item_prev contains the ID of the preceding document (e.g. the parent comment)
- feature contains a category  (e.g. the topic of the comment)
- weight contains the prevalence of the category in the range 0 to 1.
  weight Will be initialized with 1 if missing.
Items with multiple features should occur multiple times in the data table.
TODO: add weight option (instead of distinct item-feature-combinations)}
}
\value{
A tibble
}
\description{
Count sequences of features from sequences of items
}
